+++ riddle-image/src/image_ext/image.rs
@@ -1,3 +1,6 @@
+use ::image::DynamicImage;
+
 pub trait ImageImageExt {
+    fn image_from_dynimage(img: DynamicImage) -> Self;
     fn image_rgbaimage(&self) -> &::image::RgbaImage;
 }